Implement onChange prop to enable persistence
onChange allows the user of the SplitPane to receive a callback invocation
when the splitter's position is changed.  This can be persisted somewhere
like localStorage.  This commit updates README.md to include an example
of using localStorage to this end, as well as some advice on why one might
want to use something other than localStorage.