extension-of-maxSize-param-behavior
Especially if you wrap the SplitPane into a container component then
you'll need to limit the movement of the splitter (resizer) at the end
of the SplitPane (otherwise it can be dragged outside the SplitPane and
you don't catch it never more). For this purpose use the maxSize
parameter with value 0. When dragged the splitter/resizer will stop at
the border of the SplitPane component and thank this you'll be able to
pick it again and drag it back then.
And more: if you set the maxSize to negative value (e.g. -200), then the
splitter stops 200px before the border (in other words it sets the
minimal size of the 'resizable' pane in this case). This can be useful
also in the full-screen case of use.