Fix layout issues with SplitPane in WebKit browsers
It does this by doing three things:

- Uses react-prefixer to automatically prefix inline styles.
- Runs AutoPrefixer on existing external stylesheets.
- Rebuilds the demo pages.

Webkit was explored as a potential solution to this problem, but this
commit does not setup webpack.

This is primarily because the tests use Mochify to run Browserify, Mocha
and PhantomJS. To get this working with webpack, it requires some additional
refactoring of tests, which is potentially a bit disruptive when we just
want to get things rendering correctly.

Test-information:

Ran 'npm test' - all 29 tests passed successfuly.

Tested using Midori (WebKit browser), Chrome and Firefox.

Previously, the SplitPane would not render in Midori, now it renders
correctly, and can be dragged to resize panes.