feat: Support the CSS order property
Currently, if the ordering of the flex items is altered via the CSS `order` property so that the two panes are reversed, the sizing happens in the wrong direction i.e. dragging left should make the pane bigger, but it makes it smaller. This simply looks up and compares the two `order` values for each pane and inverts the `sizeDelta` to take this into account.

*Use case*: I have an iframe within a SplitPane that I need to ensure stays mounted at all times, so reversing the children is not an option.